+++ spec/integration/resource/api500/synergy/firmware_bundle/create_spec.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -9,9 +9,13 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-require_relative '../_client' # Gives access to @client
+require 'spec_helper'
 
-# NOTE: This will upload a firmware bundle
+klass = OneviewSDK::API500::Synergy::FirmwareBundle
+RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
+  let(:current_client) { $client_500_synergy }
+  let(:bundle_path) { $secrets_synergy['bundle_path'] }
+  let(:spp_path) { $secrets_synergy['spp_path'] }
 
-item = OneviewSDK::FirmwareBundle.add(@client, @firmware_bundle_path)
-puts "\nFirmware bundle #{item['name']} was successfully added."
+  include_examples 'FirmwareBundleCreateExample', 'integration api500 context'
+end